+++ tools/src/librlink/RlinkCommand.hpp
@@ -1,6 +1,6 @@
-// $Id: RlinkCommand.hpp 1092 2018-12-24 08:01:50Z mueller $
+// $Id: RlinkCommand.hpp 1121 2019-03-11 08:59:12Z mueller $
 //
-// Copyright 2011-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2011-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2019-03-10  1121   1.4.3  add BlockDoneAll()
 // 2018-12-24  1092   1.4.2  rename IsBlockExt -> HasBlockExt
 // 2018-12-23  1091   1.4.1  CmdWblk(),SetBlockWrite(): add move version
 // 2018-12-01  1076   1.4    use unique_ptr
@@ -29,7 +30,6 @@
 
 
 /*!
-  \file
   \brief   Declaration of class RlinkCommand.
 */
 
@@ -102,6 +102,7 @@ namespace Retro {
       const uint16_t*  BlockPointer() const;
       size_t        BlockSize() const;
       size_t        BlockDone() const;
+      bool          BlockDoneAll() const;
       uint8_t       Status() const;
       uint32_t      Flags() const;
       bool          TestFlagAny(uint32_t mask) const;